+++ sdk/storagecache/arm-storagecache/CHANGELOG.md
@@ -1,15 +1,45 @@
 # Release History
+    
+## 6.0.0 (2023-02-10)
+    
+**Features**
 
-## 5.2.2 (Unreleased)
+  - Added operation Caches.beginUpdate
+  - Added operation Caches.beginUpdateAndWait
+  - Added operation StorageTargets.beginRestoreDefaults
+  - Added operation StorageTargets.beginRestoreDefaultsAndWait
+  - Added Interface CachesDebugInfoHeaders
+  - Added Interface CachesDeleteHeaders
+  - Added Interface CachesFlushHeaders
+  - Added Interface CachesStartHeaders
+  - Added Interface CachesStopHeaders
+  - Added Interface CachesUpdateHeaders
+  - Added Interface CachesUpgradeFirmwareHeaders
+  - Added Interface StorageTargetFlushHeaders
+  - Added Interface StorageTargetInvalidateHeaders
+  - Added Interface StorageTargetResumeHeaders
+  - Added Interface StorageTargetsDeleteHeaders
+  - Added Interface StorageTargetsDnsRefreshHeaders
+  - Added Interface StorageTargetsRestoreDefaultsHeaders
+  - Added Interface StorageTargetsRestoreDefaultsOptionalParams
+  - Added Interface StorageTargetSuspendHeaders
+  - Interface BlobNfsTarget has a new optional parameter verificationTimer
+  - Interface BlobNfsTarget has a new optional parameter writeBackTimer
+  - Interface CachesUpdateOptionalParams has a new optional parameter resumeFrom
+  - Interface CachesUpdateOptionalParams has a new optional parameter updateIntervalInMs
+  - Interface Nfs3Target has a new optional parameter verificationTimer
+  - Interface Nfs3Target has a new optional parameter writeBackTimer
 
-### Features Added
+**Breaking Changes**
 
-### Breaking Changes
-
-### Bugs Fixed
-
-### Other Changes
-
+  - Removed operation Caches.update
+  - Operation Caches.beginCreateOrUpdate has a new signature
+  - Operation Caches.beginCreateOrUpdateAndWait has a new signature
+  - Operation StorageTargets.beginCreateOrUpdate has a new signature
+  - Operation StorageTargets.beginCreateOrUpdateAndWait has a new signature
+  - Interface CachesCreateOrUpdateOptionalParams no longer has parameter cache
+  - Interface StorageTargetsCreateOrUpdateOptionalParams no longer has parameter storagetarget
+    
 ## 5.2.1 (2023-02-02)
 
 **Features**